+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dovenv/core
 
+## 1.5.21
+
+### Patch Changes
+
+- add custom predocs supports in theme-pigeonposse
+
+- Updated dependencies []:
+  - @dovenv/utils@1.5.21
+
 ## 1.5.20
 
 ### Patch Changes